Changed science_data_command buffer interpreation to not slice off "junk" until it has received enough bytes to perceive that there is not a valid end byte.
Now the buffer accumulates bytes until it has a long enough buffer to find a possible end byte. If it is an end byte, the internal data is the message.

If it is NOT an end byte, the interpreter slices off (start byte, id byte, length) and feeds the rest into the buffer (to find the first next packet).

Added and adjusted tests to the new behavior.