Removed cobs encoding.
Since we don't have control over the packets that are created by the SIP, we can't cobs encode the whole of our custom packets. This mean the next best we can do is cobs encode only the payload of our custom packets - the byte we want to escape can still appear in the header.

Instead of cobs encoding, we will make sure that we identify and interpret packets using the header - such that the whole header needs to be corrupted, and corrupted in a specific way, before we mistakenly identify a corrupted packet as a good one.